Refactor signal handling and update types for improved performance
- Changed `rz_channel_t` from `int32_t` to `int64_t` for better compatibility with larger channel identifiers.
- Updated `rz_box` structure to use `int64_t` for integer values, enhancing the range of boxed integers.
- Modified signal management functions to accommodate the new integer type, ensuring consistency across the runtime.
- Improved debug print functions to reflect changes in data types.
- Adjusted tests to use `console_out_signal` instead of `output_int_signal` for better clarity and consistency.
- Added end-to-end tests to validate console output functionality.
- Introduced a test for annotated polymorphic top-level instantiation to ensure type correctness.

+++ src/runtime/channel.h
@@ -3,7 +3,7 @@
 #include "stdint.h"
 #include "core.h"
 
-typedef int32_t rz_channel_t;
+typedef int64_t rz_channel_t;
 
 static const rz_channel_t RZ_CHANNEL_CONSOLE_IN = 0;
